feat:create forecast details page, show details
